refactor(antd): one file per component, matching mantine
chrome.tsx split into Toolbar, ActiveFilterChips, BulkActionBar, ErrorState
and FilterDrawer. The adapter drives antd's own high-level Table, so it has
no tables.tsx to split. No behaviour, exports or public API change.
